% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_dep_add.R
\name{renv_dep_add}
\alias{renv_dep_add}
\title{Add Packages to \verb{_dependencies.R} for renv}
\usage{
renv_dep_add(pkg)
}
\arguments{
\item{pkg}{Character vector of package names.
Packages to add to \verb{_dependencies.R} and install if not already installed.
Can use the "\verb{<remote>/<repo>}" syntax for installing from remotes (e.g., GitHub), provided that \verb{<repo>} is also the name of the package.}
}
\value{
Invisibly returns \code{TRUE} upon successful completion.
}
\description{
Adds \verb{library(<pkg>)} calls for each package listed to the \verb{_dependencies.R} file.
Installs packages if they are not already installed.
}
\examples{
\dontrun{
# Add and install CRAN packages
renv_dep_add(c("dplyr", "ggplot2"))

# Add and install a GitHub package
renv_dep_add("hadley/httr")
}

}
