% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_dep_add.R
\name{renvvv_dep_add}
\alias{renvvv_dep_add}
\title{Add Packages to \verb{_dependencies.R} for renv}
\usage{
renvvv_dep_add(pkg, force = FALSE)
}
\arguments{
\item{pkg}{Character vector of package names.
Packages to add to \verb{_dependencies.R} and install if not already installed.
Can use the "\verb{<remote>/<repo>}" syntax for installing from remotes
(e.g., GitHub), provided that \verb{<repo>} is also the name of the package.}

\item{force}{Logical. If \code{FALSE} (default), prompts for confirmation in
interactive sessions or errors in non-interactive sessions. Set to \code{TRUE}
to proceed without prompting.}
}
\value{
Invisibly returns \code{TRUE} upon successful completion.
}
\description{
Adds \verb{library(<pkg>)} calls for each package listed to the
\verb{_dependencies.R} file.
Installs packages if they are not already installed.
}
\examples{
\dontrun{
# Add and install CRAN packages (interactive mode will prompt)
renvvv_dep_add(c("dplyr", "ggplot2"))

# Add and install a GitHub package
renvvv_dep_add("hadley/httr")

# Non-interactive mode requires force = TRUE
renvvv_dep_add(c("dplyr", "ggplot2"), force = TRUE)
}

}
