% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{renvvv_install}
\alias{renvvv_install}
\alias{renv3_install}
\title{Install Packages Robustly}
\usage{
renvvv_install(packages, prompt = FALSE, args_install = list())
}
\arguments{
\item{packages}{A character vector of packages to install. Can contain CRAN,
Bioconductor, or GitHub package references (e.g., \code{"dplyr"}, \code{"user/repo@v1"}).}

\item{prompt}{Logical. Whether to prompt for user confirmation during
installation. Default is \code{FALSE}.}

\item{args_install}{List. Arbitrary arguments to pass down to \code{renv::install()}.}
}
\value{
Invisibly returns \code{TRUE} upon completion.
}
\description{
A "tryhard" version of \code{renv::install()}. It first attempts to install all
specified packages in a single vectorized call. If that fails, it compares
the current installed packages and their versions against the state prior
to the installation attempt. Any package that is still missing or whose
version did not change will then be installed individually, ensuring that
one failing package does not prevent the installation of others.
}
\examples{
\dontrun{
# Install multiple packages robustly
renvvv_install(c("dplyr", "ggplot2", "SATVILab/projr@dev"))

# Pass arbitrary parameters down to renv::install
renvvv_install(c("dplyr"), args_install = list(rebuild = TRUE))
}

}
