% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{renvvv_renv_repos_setup}
\alias{renvvv_renv_repos_setup}
\title{Make .Rprofile source script to make renv use right repos}
\usage{
renvvv_renv_repos_setup()
}
\description{
Copied from
\url{https://github.com/rstudio/renv/issues/1052#issuecomment-1342567839}.
}
