% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{setup_hpc_renv}
\alias{setup_hpc_renv}
\title{Make .Rprofile source script to make renv use scratch directory}
\usage{
setup_hpc_renv()
}
\description{
Make .Rprofile source script to make renv use scratch directory
}
